pyproject.toml
[project]
name = "search-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.11.0",
    "litellm>=1.75.8",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "tavily-python>=0.7.11",
]
